heart_mirror/requirements.txt
Flask==2.3.3
Flask-CORS==4.0.0
Flask-JWT-Extended==4.7.1
Werkzeug==3.1.8
PyJWT==2.12.1
gunicorn==21.2.0
bcrypt==4.0.1
python-dotenv==1.0.0